+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Client timeout in CLI.
 - Support for header parameters in operations.
 - Validation for non-body parameters.
+- `x-rust-dependencies` field in root for specifying additional crate dependencies.
+- Default en/decoders (JSON/YAML) and support for custom en/decoders through `x-rust-coders` field in root.
 
 ### Changed
 - Switched to templating for (almost) static modules.